# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GenerateHtmlInput {
  prompt: string;
  username?: string;
}

export interface GenerateHtmlResult {
  html: string;
  title: string;
}

export interface ChatMessage {
  id: string;
  username: string;
  text: string;
  timestamp: string;
  initials: string;
}

export interface ChatMessageInput {
  username: string;
  text: string;
}

export interface OnlineUser {
  username: string;
  initials: string;
  lastSeen: string;
}