# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: postChatMessage
      tags: [chat]
      summary: Post a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        text:
          type: string
        timestamp:
          type: string
          format: date-time
        initials:
          type: string
      required:
        - id
        - username
        - text
        - timestamp
        - initials
    ChatMessageInput:
      type: object
      properties:
        username:
          type: string
        text:
          type: string
      required:
        - username
        - text
...